fix(preview): match TP talk fade and do_heightcenter layout
Talk windows used an opaque bar and packed lines_per_page into the text box, so two-line dialogue sat at the top with static spacing. Use the I8 112/8_01 fade (HIO alpha 0.9), BLO mg_e4lin fontSize/lineSpace, and center the used lines like jmessage_tRenderingProcessor::do_heightcenter.

+++ core/bfn_core.py
@@ -414,7 +414,7 @@ class BfnCore:
 
         return self._qimages_cache
 
-    def layout_text(self, text: str, translation_map: Optional[Dict[str, str]] = None, line_spacing: int = 10,
+    def layout_text(self, text: str, translation_map: Optional[Dict[str, str]] = None, line_spacing: float = 10,
                     char_spacing: float = 0, colors: Optional[List[Optional[str]]] = None,
                     scales: Optional[List[float]] = None,
                     icons: Optional[Dict[int, Dict[str, Any]]] = None) -> Tuple[List[Dict[str, Any]], int, int]:
@@ -449,9 +449,9 @@ class BfnCore:
             - total height of text block
         """
         try:
-            line_spacing = int(line_spacing)
+            line_spacing = float(line_spacing)
         except (TypeError, ValueError):
-            line_spacing = 10
+            line_spacing = 10.0
         try:
             char_spacing = float(char_spacing)
         except (TypeError, ValueError):
